chore: Refactor websocket structure
