+++ package.json
@@ -65,7 +65,6 @@
     "rollup-plugin-typescript2": "0.29.0",
     "rollup-plugin-url": "3.0.1",
     "ts-jest": "26.4.4",
-    "typescript": "4.1.3",
-    "md5": "2.3.0"
+    "typescript": "4.1.3"
   }
 }
