+++ packages/iqoption/lib/websocket/events/Heartbeat.ts
@@ -1,0 +1,15 @@
+import { WebSocketEvent } from '../../types'
+import { BaseEventHandler } from '../BaseEventHandler'
+
+export class Heartbeat extends BaseEventHandler<number> {
+  public get name(): string {
+    return 'heartbeat'
+  }
+
+  public handle(event: WebSocketEvent<number>): void {
+    this.webSocket.send('heartbeat', {
+      heartbeatTime: String(event.msg),
+      userTime: String(Date.now()),
+    })
+  }
+}
